feat(thales): Arbitrum Vaults (#2517)
